refactor: inline campaign failure handling
